Fixes #625. I added command line options for the maximum depth of the ancestors and descendants, and also the format of the output. The user can now set the output directly to pdf, for instance, and doesn't have to convert himself. I also moved the functionality for the drawing to aiida.common.graph, so that it can be used also from the shell and not only by command line
